refactor: `CreateNode`
