+++ Backend/src/Modules/Nodes/Nodes.Api/Controllers/Nodes/NodesResponses.cs
@@ -1,0 +1,5 @@
+﻿using BuildingBlocks.Domain.Nodes.Node.ValueObjects;
+
+namespace Nodes.Api.Controllers.Nodes;
+
+public record CreateNodeResponse(NodeId Id);
