+++ docker-compose/dev/docker-compose.override.yml
@@ -11,17 +11,17 @@ services:
       - "26000:8080"
       - "26061:8081"
 
-  tl-frontend:
-    container_name: tl-frontend
-    image: ${DOCKER_REGISTRY-}tl-frontend
-    networks:
-      - main
-    depends_on:
-      - tl-api
-    build:
-      context: ../Frontend/react-app
-    ports:
-      - "24000:23000"
+  # tl-frontend:
+  #   container_name: tl-frontend
+  #   image: ${DOCKER_REGISTRY-}tl-frontend
+  #   networks:
+  #     - main
+  #   depends_on:
+  #     - tl-api
+  #   build:
+  #     context: ../Frontend/react-app
+  #   ports:
+  #     - "24000:23000"
 
   tl-database:
     environment:
@@ -34,8 +34,8 @@ services:
     volumes:
       - timelines_core:/var/lib/postgresql/data
 
-  proxy:
-    depends_on:
-      - tl-frontend
-    volumes:
-      - ${PWD}/nginx.conf:/etc/nginx/conf.d/default.conf:ro
+  # proxy:
+  #   depends_on:
+  #     - tl-frontend
+  #   volumes:
+  #     - ${PWD}/nginx.conf:/etc/nginx/conf.d/default.conf:ro
